+++ docs/changelog.md
@@ -1,5 +1,6 @@
 2025/04/xx Release 4.0.2
 
+- matchbox validation: html tags in result [#371](https://github.com/ahdis/matchbox/issues/371)
 - matchbox validation: make showMessagesFromReferences default to true [#370](https://github.com/ahdis/matchbox/issues/370)
 
 2025/04/15 Release 4.0.1